Export SerializeSCTList and DeserializeSCTList and move to helpers/helpers.go
